Add authentication

+++ db/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from schemas import UserBase
 from db.models import DbUser
 from db.hash import Hash
+from fastapi import HTTPException, status
 
 def create_user(db: Session, request: UserBase):
     new_user = DbUser(
@@ -18,13 +19,25 @@ def create_user(db: Session, request: UserBase):
 def get_all_users(db: Session):
     return db.query(DbUser).all()
 
-def get_user(db: Session, id: int):
-    # Handle any exceptions
-    return db.query(DbUser).filter(DbUser.id == id).first()
+def get_user(db: Session, id: int):   
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id '{id}' not found")
+    return user
+        
+def get_user_by_username(db: Session, username: str):   
+    user = db.query(DbUser).filter(DbUser.username == username).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with useranme '{username}' not found")
+    return user
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
-    # Handle any exceptions
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id '{id}' not found")
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -35,7 +48,9 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
-    # Handle any exceptions
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id '{id}' not found")
     db.delete(user)
     db.commit()
     return 'ok'
